+++ src/lighthouse/lighthouse.ts
@@ -1,28 +1,10 @@
 import { Options } from "../types";
 
 import { runLighthouseAndGetReports } from "./runner";
+import { LighthouseReport } from "./types";
 
 export async function getLighthouseReport(options: Options): Promise<LighthouseReport> {
   const lighthouseReport = await runLighthouseAndGetReports(options.url);
 
-  const metrics = lighthouseReport.categoryReport;
-  const { htmlReport } = lighthouseReport;
-
-  return {
-    htmlReport,
-    metrics,
-  };
-}
-
-export interface LighthouseMetrics {
-  performance?: number;
-  accessibility?: number;
-  "best-practices"?: number;
-  seo?: number;
-  pwa?: number;
-}
-
-export interface LighthouseReport {
-  htmlReport: string;
-  metrics: LighthouseMetrics;
+  return lighthouseReport;
 }
